Fix: ID 타입 수정
- Repository ID 타입 Long 타입으로 수정

+++ src/main/java/ic2/onboarding/survey/repository/SurveyRepository.java
@@ -3,5 +3,5 @@ package ic2.onboarding.survey.repository;
 import ic2.onboarding.survey.entity.Survey;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SurveyRepository extends JpaRepository<Survey, Integer> {
+public interface SurveyRepository extends JpaRepository<Survey, Long> {
 }
